Re-do web connectors with placement functions instead of posts
This isn't necessarily better/easier, but more flexible because e.g. arbitrary
polyhedrons can be added to smooth geometry. It also results in meshes with
less vertices (one per corner, not three or four because of the posts).

The web connectors are still manifold - at least until the plates are unioned
in! It seems OpenSCAD isn't capable of generating manifold meshes unless the
coordinates are *exactly* the same. I'm exploring if it's e.g. possible to
export the plates separately from the webs and walls, and them combine them
in the slicer or something. In any case, hopefully the simplified geometry
slices better even if it is non-manifold. We'll see once the walls are done.