Re-do walls with position functions instead of posts
I think this is easier than using the posts, for several reasons. The mesh has
fewer vertices, and when unioned the mesh is manifold (except when the key and
thumb plates are unioned in).

It will also make placing other objects like the USB holder and screw holes much
easier. Since the exact coordinates of the edge points is known, they can easily
be projected onto the base plane (in Clojure) to get a point for translation.

To me, storing the edges in hash-maps with keys/names also makes reasoning about
the geometry more straight-forward, same with the directions.

Finally, the vertical alignment of plate edges produces more consistent walls.
The chamfer is the same angle everywhere, and the walls have a uniform
thickness.

Overall, I'm happy with this approach. Having to derive the position functions
is tedious, and has to be kept in sync with the plate placement functions.
Polyhedrons are slightly tricky, since faces have to be defined clockwise. But
they are a lot more powerful and produce simpler geometry than hulls. So the
benefits to web, wall, and other geometry construction is huge.